adição de componente de input de texto e telas de onboarding, e nova organização das pastas
